Fix #62 Update the prefix for /procurement-buyer/verify-certificate to /verifications/verify-certificate
