+++ buyer/urls.py
@@ -7,5 +7,4 @@ urlpatterns = [
     path('tender/<int:tender_id>/', views.get_tender),
     path('tender/<int:tender_id>/publish-tender', views.publish_tender),
     path('qualification-documents',views.get_qualification_documents),
-    path('verify-certificate', views.verify_certificate),
 ]
