+++ pob_backend/urls.py
@@ -24,5 +24,6 @@ urlpatterns = [
     path(r"certificates", include("certificate.urls")),
     path(r'procurement-buyer/', include("buyer.urls")),
     path(r'procurement-supplier/', include("seller.urls")),
+    path(r'verifications/', include("verifications.urls")),
     path(r'webhooks/', include("webhook.urls")),
 ]
