+++ verifications/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('verify-certificate', views.verify_certificate),
+]
